# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"DKBlytics: a modular API to your DKB banking","description":"A simple API to track personal banking transactions and balances.","version":"0.1.0"},"paths":{"/api/accounts/":{"post":{"tags":["Accounts"],"summary":"Create Account","description":"Create a new account.\n\nReturns 201 on success.\nReturns 409 if an account with the same name already exists.","operationId":"create_account_api_accounts__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Accounts"],"summary":"Get All Accounts","description":"Retrieve all accounts (sorted by name).","operationId":"get_all_accounts_api_accounts__get","parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by account name (exact).","title":"Name"},"description":"Filter by account name (exact)."},{"name":"holder","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by holder_name (exact).","title":"Holder"},"description":"Filter by holder_name (exact)."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"},"title":"Response Get All Accounts Api Accounts  Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/accounts/{public_id}":{"put":{"tags":["Accounts"],"summary":"Update Account","description":"Update an existing account's balance.\n\n- The body `name` must match the path `public_id`.\n- Returns 404 if the account does not exist.\n- Returns 400 if the names do not match.\n- Returns 409 if the IBAN is already taken by another account.","operationId":"update_account_api_accounts__public_id__put","parameters":[{"name":"public_id","in":"path","required":true,"schema":{"type":"string","title":"Public Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Accounts"],"summary":"Get Account","description":"Retrieve a specific account by public ID.","operationId":"get_account_api_accounts__public_id__get","parameters":[{"name":"public_id","in":"path","required":true,"schema":{"type":"string","title":"Public Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/balances/series":{"get":{"tags":["Balances"],"summary":"Get Balance Series","description":"Return a lightweight series with `date` + `balance` at the chosen granularity.","operationId":"get_balance_series_api_balances_series_get","parameters":[{"name":"account_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Optional account filter.","title":"Account Id"},"description":"Optional account filter."},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Inclusive YYYY-MM-DD.","title":"Date From"},"description":"Inclusive YYYY-MM-DD."},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Inclusive YYYY-MM-DD.","title":"Date To"},"description":"Inclusive YYYY-MM-DD."},{"name":"granularity","in":"query","required":false,"schema":{"$ref":"#/components/schemas/Granularity","description":"Bucket size for the returned series (daily/weekly/monthly/fiscal_monthly/yearly).","default":"daily"},"description":"Bucket size for the returned series (daily/weekly/monthly/fiscal_monthly/yearly)."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BalancePoint"},"title":"Response Get Balance Series Api Balances Series Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/balances/surplus":{"get":{"tags":["Balances"],"summary":"Get Surplus Series","description":"Return aggregated deltas (\"surplus\") for the given range.","operationId":"get_surplus_series_api_balances_surplus_get","parameters":[{"name":"account_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Optional account filter.","title":"Account Id"},"description":"Optional account filter."},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Inclusive YYYY-MM-DD.","title":"Date From"},"description":"Inclusive YYYY-MM-DD."},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Inclusive YYYY-MM-DD.","title":"Date To"},"description":"Inclusive YYYY-MM-DD."},{"name":"granularity","in":"query","required":false,"schema":{"$ref":"#/components/schemas/Granularity","description":"Bucket size for the returned surplus buckets (daily/weekly/monthly/fiscal_monthly/yearly).","default":"daily"},"description":"Bucket size for the returned surplus buckets (daily/weekly/monthly/fiscal_monthly/yearly)."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SurplusPoint"},"title":"Response Get Surplus Series Api Balances Surplus Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/transactions/":{"post":{"tags":["Transactions"],"summary":"Create Transaction","description":"Create a transaction.\n\n- Duplicate policy:\n  * Allowed: duplicates when SAME (fingerprint, batch_hash)\n  * Rejected (409): fingerprint exists with a DIFFERENT batch_hash\n- Category is resolved from rules for the response.","operationId":"create_transaction_api_transactions__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Transactions"],"summary":"Get All Transactions","description":"List transactions with pagination, sorting, and filters.","operationId":"get_all_transactions_api_transactions__get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"exclusiveMinimum":0,"description":"Page size.","default":50,"title":"Limit"},"description":"Page size."},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"description":"Zero-based start index.","default":0,"title":"Offset"},"description":"Zero-based start index."},{"name":"sort_by","in":"query","required":false,"schema":{"type":"string","pattern":"^(date_desc|date_asc|amount_desc|amount_asc)$","description":"Sort order.","default":"date_desc","title":"Sort By"},"description":"Sort order."},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Inclusive YYYY-MM-DD.","title":"Date From"},"description":"Inclusive YYYY-MM-DD."},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Inclusive YYYY-MM-DD.","title":"Date To"},"description":"Inclusive YYYY-MM-DD."},{"name":"account_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by account public_id.","title":"Account Id"},"description":"Filter by account public_id."},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Optional category filter.\n- Leave empty to list all transactions.\n- Pass a category name to filter by that category.\n- Pass 'null' to list uncategorized transactions.","title":"Category"},"description":"Optional category filter.\n- Leave empty to list all transactions.\n- Pass a category name to filter by that category.\n- Pass 'null' to list uncategorized transactions."},{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Case-insensitive search in entity/text/reference.","title":"Q"},"description":"Case-insensitive search in entity/text/reference."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedTransactions"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/transactions/summary":{"get":{"tags":["Transactions"],"summary":"Summarize Transactions By Category","description":"Summarize transactions **by category nodes** at a given depth within an optional scope.\n\nExamples:\n- `scope_name=None, depth=1` ⇒ group by root categories.\n- `scope_name=\"Expenses\", depth=1` ⇒ group by direct children of “Expenses”.\n- `scope_name=\"Expenses\", depth=2` ⇒ group by grandchildren of “Expenses”.\n\nNotes:\n- If a resolved transaction category is shallower than the requested depth within the scope,\n  it is grouped under its deepest available ancestor (not dropped).\n- You can filter via `account`, `date_from`, `date_to`, and `q`. There is **no** separate “group by account”.","operationId":"summarize_transactions_by_category_api_transactions_summary_get","parameters":[{"name":"scope_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Parent category name whose subtree is the scope. Omit for global roots.","title":"Scope Name"},"description":"Parent category name whose subtree is the scope. Omit for global roots."},{"name":"depth","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Depth within the scope to define the groups. 1 = direct children; 2 = grandchildren; etc.","default":1,"title":"Depth"},"description":"Depth within the scope to define the groups. 1 = direct children; 2 = grandchildren; etc."},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Inclusive YYYY-MM-DD.","title":"Date From"},"description":"Inclusive YYYY-MM-DD."},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Inclusive YYYY-MM-DD.","title":"Date To"},"description":"Inclusive YYYY-MM-DD."},{"name":"account_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by account public_id.","title":"Account Id"},"description":"Filter by account public_id."},{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Case-insensitive search in entity/text/reference.","title":"Q"},"description":"Case-insensitive search in entity/text/reference."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionSummary"},"title":"Response Summarize Transactions By Category Api Transactions Summary Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/transactions/{tx_id}":{"get":{"tags":["Transactions"],"summary":"Get Transaction","description":"Retrieve a single transaction.","operationId":"get_transaction_api_transactions__tx_id__get","parameters":[{"name":"tx_id","in":"path","required":true,"schema":{"type":"integer","title":"Tx Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Transactions"],"summary":"Update Transaction","description":"Update transaction entity and/or text fields.\n\nThis endpoint allows manual correction of generic transaction descriptions\nwhile preserving the original fingerprint for deduplication integrity.\nOnly the entity and text fields can be modified.","operationId":"update_transaction_api_transactions__tx_id__put","parameters":[{"name":"tx_id","in":"path","required":true,"schema":{"type":"integer","title":"Tx Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/categories/":{"get":{"tags":["Categories"],"summary":"Get All Categories","description":"Retrieve all existing categories.","operationId":"get_all_categories_api_categories__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Category"},"type":"array","title":"Response Get All Categories Api Categories  Get"}}}},"404":{"description":"Not found"}}},"post":{"tags":["Categories"],"summary":"Create Category","description":"Add a new category.","operationId":"create_category_api_categories__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/categories/tree":{"get":{"tags":["Categories"],"summary":"Get Category Tree","description":"Return a nested category tree.\n\n- No name -> all root categories (no parent) with full subtrees.\n- With name -> only that category's direct children (each with its subtree).","operationId":"get_category_tree_api_categories_tree_get","parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"If provided, return only the children (with subtrees) of this category name.","title":"Name"},"description":"If provided, return only the children (with subtrees) of this category name."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Get Category Tree Api Categories Tree Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/categories/{name}":{"get":{"tags":["Categories"],"summary":"Get Category By Name","description":"Retrieve a category by its name.\n\nNote: If multiple categories share this name under different parents,\nthe request is considered ambiguous and returns 409.","operationId":"get_category_by_name_api_categories__name__get","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Categories"],"summary":"Delete Category","description":"Delete a category by name.","operationId":"delete_category_api_categories__name__delete","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/rules/":{"get":{"tags":["Category Rules"],"summary":"Get All Category Rules","description":"List all category rules.","operationId":"get_all_category_rules_api_rules__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CategoryRule"},"type":"array","title":"Response Get All Category Rules Api Rules  Get"}}}},"404":{"description":"Not found"}}},"post":{"tags":["Category Rules"],"summary":"Create Category Rule","description":"Create a category rule (general or transaction-specific).\n\nRule hierarchy:\n  1) transaction-specific rule (transaction_id match)\n  2) exact match on (entity AND text)  \n  3) default match on (entity AND text IS NULL)\n  \nAfter creating the rule, automatically recalculates matching transactions only.","operationId":"create_category_rule_api_rules__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryRuleCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryRule"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/rules/{rule_id}":{"delete":{"tags":["Category Rules"],"summary":"Delete Category Rule","description":"Delete a category rule by id.\n\nAfter deleting the rule, recalculates only the transactions that could be\naffected by the removed scope.","operationId":"delete_category_rule_api_rules__rule_id__delete","parameters":[{"name":"rule_id","in":"path","required":true,"schema":{"type":"integer","title":"Rule Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/rules/resolve":{"get":{"tags":["Category Rules"],"summary":"Resolve Rule","description":"Resolve a category name for a given (entity, text).\n\nNote: This endpoint only resolves general rules (entity/text based).\nFor transaction-specific rules, use the transaction endpoints directly.\n\nResolution order:\n  1) exact (entity AND text)\n  2) default (entity AND text IS NULL)\nReturns the category name or null if no match.","operationId":"resolve_rule_api_rules_resolve_get","parameters":[{"name":"entity","in":"query","required":true,"schema":{"type":"string","description":"Transaction entity","title":"Entity"},"description":"Transaction entity"},{"name":"text","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Transaction text/description","title":"Text"},"description":"Transaction text/description"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Response Resolve Rule Api Rules Resolve Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/rules/apply":{"post":{"tags":["Category Rules"],"summary":"Apply Rules To Transactions","description":"Recalculate categories for a subset of transactions based on current rules.\nWhen no filters are provided this still updates all transactions.","operationId":"apply_rules_to_transactions_api_rules_apply_post","parameters":[{"name":"transaction_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Limit recalculation to a specific transaction.","title":"Transaction Id"},"description":"Limit recalculation to a specific transaction."},{"name":"entity","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Limit recalculation to transactions matching this entity.","title":"Entity"},"description":"Limit recalculation to transactions matching this entity."},{"name":"text","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"When entity is provided, further filter by description/text.","title":"Text"},"description":"When entity is provided, further filter by description/text."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/bank/update_from_bank":{"post":{"tags":["Bank Integration"],"summary":"Update From Bank","description":"Connect to the bank API, fetch new transactions, insert them,\nand update account balances.\n\nReturns a mapping of account name -> number of inserted transactions for this run.","operationId":"update_from_bank_api_bank_update_from_bank_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"Not found"}}}}},"components":{"schemas":{"Account":{"properties":{"id":{"type":"integer","title":"Id","description":"Database row ID."},"public_id":{"type":"string","title":"Public Id","description":"Opaque, stable identifier to reference this account in APIs."},"name":{"type":"string","maxLength":255,"title":"Name","description":"Human-friendly account name."},"holder_name":{"type":"string","maxLength":255,"title":"Holder Name","description":"Account holder's full name."},"iban_last4":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Iban Last4","description":"Last 4 characters of the IBAN for display; may be null if not available."},"balance":{"type":"string","title":"Balance","description":"Current balance."}},"additionalProperties":false,"type":"object","required":["id","public_id","name","holder_name","balance"],"title":"Account","description":"Account as returned by the API (IBAN is not exposed)."},"AccountCreate":{"properties":{"name":{"type":"string","maxLength":255,"title":"Name","description":"Human-friendly account name (not unique)."},"holder_name":{"type":"string","maxLength":255,"title":"Holder Name","description":"Account holder's full name as provided by the bank."},"iban_plain":{"type":"string","maxLength":34,"title":"Iban Plain","description":"Raw IBAN. Used server-side only to compute a keyed digest; never returned."},"balance":{"anyOf":[{"type":"number","minimum":0.0},{"type":"string"}],"title":"Balance","description":"Current balance for this account."}},"additionalProperties":false,"type":"object","required":["name","holder_name","iban_plain","balance"],"title":"AccountCreate","description":"Payload to create or update an account (IBAN never leaves the API)."},"BalancePoint":{"properties":{"date":{"type":"string","format":"date","title":"Date","description":"Representative date for the data point."},"balance":{"type":"string","title":"Balance","description":"Closing balance on the given date."}},"additionalProperties":false,"type":"object","required":["date","balance"],"title":"BalancePoint","description":"Minimal time-series point for balances."},"Category":{"properties":{"id":{"type":"integer","title":"Id","description":"Category ID."},"name":{"type":"string","maxLength":255,"title":"Name","description":"Category name."},"parent_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Name","description":"Direct parent category's name, if any."}},"additionalProperties":false,"type":"object","required":["id","name"],"title":"Category","description":"Category as returned by the API."},"CategoryCreate":{"properties":{"name":{"type":"string","maxLength":255,"title":"Name","description":"Category name."},"parent_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Name","description":"Direct parent category's name; omit for a root category."}},"additionalProperties":false,"type":"object","required":["name"],"title":"CategoryCreate","description":"Payload to create a transaction category."},"CategoryRule":{"properties":{"id":{"type":"integer","title":"Id","description":"CategoryRule ID."},"transaction_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Transaction Id","description":"Transaction ID for transaction-specific rules; null means general rule."},"text":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Text","description":"Exact description text to match; null means entity-wide default."},"entity":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Entity","description":"Exact counterparty/entity to match."},"category_name":{"type":"string","title":"Category Name","description":"Name of the category assigned by this rule."}},"additionalProperties":false,"type":"object","required":["id","category_name"],"title":"CategoryRule","description":"Category rule as returned by the API."},"CategoryRuleCreate":{"properties":{"transaction_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Transaction Id","description":"Transaction ID for transaction-specific rules. When provided, rule applies only to this transaction."},"text":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Text","description":"Exact description text to match; use null for an entity-wide default."},"entity":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Entity","description":"Exact counterparty/entity to match. Required for non-transaction rules."},"category_name":{"type":"string","title":"Category Name","description":"Name of the category to assign when the rule matches."}},"additionalProperties":false,"type":"object","required":["category_name"],"title":"CategoryRuleCreate","description":"Payload to create a category rule.\n\nRule hierarchy (implemented in the service):\n  1) transaction-specific rule (transaction_id match)\n  2) exact match on (entity AND text)\n  3) default match on (entity AND text IS NULL)"},"Granularity":{"type":"string","enum":["daily","weekly","monthly","yearly","fiscal_monthly"],"title":"Granularity"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaginatedTransactions":{"properties":{"total":{"type":"integer","minimum":0.0,"title":"Total","description":"Total number of matching transactions."},"limit":{"type":"integer","exclusiveMinimum":0.0,"title":"Limit","description":"Page size (maximum number of items)."},"offset":{"type":"integer","minimum":0.0,"title":"Offset","description":"Zero-based index of the first item."},"items":{"items":{"$ref":"#/components/schemas/Transaction"},"type":"array","title":"Items","description":"Transactions within the current page."}},"additionalProperties":false,"type":"object","required":["total","limit","offset","items"],"title":"PaginatedTransactions","description":"Container for paginated transaction listings."},"SurplusPoint":{"properties":{"date":{"type":"string","format":"date","title":"Date","description":"Representative date for the data point."},"delta":{"type":"string","title":"Delta","description":"Net change during the bucket ending on `date`."}},"additionalProperties":false,"type":"object","required":["date","delta"],"title":"SurplusPoint","description":"Minimal time-series point for surplus/net change."},"Transaction":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id","description":"Transaction ID."},"text":{"type":"string","maxLength":1000,"title":"Text","description":"Transaction description text."},"entity":{"type":"string","maxLength":500,"title":"Entity","description":"Counterparty/entity name."},"account_id":{"type":"string","title":"Account Id","description":"Account public_id this transaction belongs to."},"account_name":{"type":"string","maxLength":255,"title":"Account Name","description":"Human-friendly account name."},"amount":{"type":"string","title":"Amount","description":"Signed amount (negative = outflow, positive = inflow)."},"date":{"type":"string","format":"date","title":"Date","description":"Booking date in YYYY-MM-DD format."},"reference":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Reference","description":"Optional customer reference."},"fingerprint":{"anyOf":[{"type":"string","maxLength":64,"pattern":"^[0-9a-f]{64}$"},{"type":"null"}],"title":"Fingerprint","description":"64-char hex fingerprint for de-duplication."},"batch_hash":{"anyOf":[{"type":"string","maxLength":64},{"type":"null"}],"title":"Batch Hash","description":"Batch identifier, if provided."},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category","description":"Resolved category name based on rules, if any."}},"additionalProperties":false,"type":"object","required":["text","entity","account_id","account_name","amount","date"],"title":"Transaction","description":"Transaction as returned by the API."},"TransactionCreate":{"properties":{"text":{"type":"string","maxLength":1000,"title":"Text","description":"Transaction description text from the statement."},"entity":{"type":"string","maxLength":255,"title":"Entity","description":"Counterparty/entity name as extracted from the statement."},"account_id":{"type":"string","title":"Account Id","description":"Account public_id this transaction belongs to."},"amount":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Amount","description":"Signed amount (negative = outflow, positive = inflow)."},"date":{"type":"string","format":"date","title":"Date","description":"Booking date in YYYY-MM-DD format."},"reference":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Reference","description":"Optional customer reference or memo."},"batch_hash":{"anyOf":[{"type":"string","maxLength":64},{"type":"null"}],"title":"Batch Hash","description":"Optional batch identifier. Duplicates are allowed within the same batch; cross-batch duplicate fingerprints are rejected."}},"additionalProperties":false,"type":"object","required":["text","entity","account_id","amount","date"],"title":"TransactionCreate","description":"Payload to create a transaction."},"TransactionSummary":{"properties":{"key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key","description":"Grouping key (e.g., category name when grouped by category). None represents 'uncategorized'."},"amount_sum":{"type":"string","title":"Amount Sum","description":"Sum of amounts in this group (negative = net outflow, positive = net inflow)."},"transactions":{"items":{"$ref":"#/components/schemas/Transaction"},"type":"array","title":"Transactions","description":"All transactions that contributed to this group's amount."}},"additionalProperties":false,"type":"object","required":["amount_sum"],"title":"TransactionSummary","description":"Aggregate summary for transactions grouped by a key."},"TransactionUpdate":{"properties":{"entity":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Entity","description":"New counterparty/entity name. Leave null to keep existing value."},"text":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Text","description":"New transaction description text. Leave null to keep existing value."}},"additionalProperties":false,"type":"object","title":"TransactionUpdate","description":"Payload to update transaction entity and/or text fields."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}